extract github release fetching
