+++ src/@types/jest-extended.d.ts
@@ -1,8 +1,9 @@
 import * as matchers from "jest-extended";
 import { jest } from "@jest/globals";
 
-declare module "expect" {
-	type JestExtendedMatchers = typeof matchers;
+type JestExtendedMatchers = typeof matchers;
+
+declare module "@jest/expect" {
 	// eslint-disable-next-line @typescript-eslint/no-empty-interface, @typescript-eslint/no-unused-vars
 	export interface Matchers<R> extends JestExtendedMatchers {
 		toHaveBeenCalledBefore(
