+++ src/test/github.test.ts
@@ -1,0 +1,8 @@
+import { describe } from "@jest/globals";
+import test from "node:test";
+
+describe("fetchLatestRelease", () => {
+	test("downloads the latest release from Github", () => {
+		// TODO
+	});
+});
