+++ src/test/release.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, test } from "@jest/globals";
-import Github from "../github";
 import Release from "../release";
 import { URI } from "vscode-uri";
 import * as semver from "semver";
+import GithubRelease from "../github/release";
 
 describe("fromGithubRelease", () => {
 	test("should throw given github release without a name", () => {
-		const githubRelease: Github.Release = {
+		const githubRelease: GithubRelease.T = {
 			name: null,
 			assets: [],
 		};
@@ -15,7 +15,7 @@ describe("fromGithubRelease", () => {
 	});
 
 	test("should throw given github release name is not valid version", () => {
-		const githubRelease: Github.Release = {
+		const githubRelease: GithubRelease.T = {
 			name: "hello",
 			assets: [],
 		};
@@ -24,7 +24,7 @@ describe("fromGithubRelease", () => {
 	});
 
 	test("should throw given github release does not have a lexical release asset", () => {
-		const githubRelease: Github.Release = {
+		const githubRelease: GithubRelease.T = {
 			name: "2023-05-27T15:48:20",
 			assets: [],
 		};
@@ -34,7 +34,7 @@ describe("fromGithubRelease", () => {
 
 	test("should create a release with a date version", () => {
 		const githubReleaseName = "2023-05-27T15:48:20";
-		const githubRelease: Github.Release = {
+		const githubRelease: GithubRelease.T = {
 			name: githubReleaseName,
 			assets: [
 				// eslint-disable-next-line @typescript-eslint/naming-convention
@@ -55,7 +55,7 @@ describe("fromGithubRelease", () => {
 
 	test("should create a release with a semantic version", () => {
 		const githubReleaseName = "v1.2.3";
-		const githubRelease: Github.Release = {
+		const githubRelease: GithubRelease.T = {
 			name: githubReleaseName,
 			assets: [
 				// eslint-disable-next-line @typescript-eslint/naming-convention
